services/rag_service/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
supabase
google-genai
google-api-core
numpy
PyMuPDF
pydub
moviepy
openai
